+++ src/Files/profile.ps1
@@ -127,14 +127,12 @@ else {
     Update-ProfileModule
 }
 
-if ($PSVersionTable.PSVersion -ge [System.Version]"7.0") {
-    if ((Get-Module "PSReadline" -ListAvailable).Version -ge [System.Version]"2.2.0") {
-        Set-PSReadLineOption -PredictionSource History
-        Set-PSReadLineOption -PredictionViewStyle ListView
-    }
-    else {
-        "Missing PSReadline 2.2.0 or newer, will not set options"
-    }
+if ((Get-Module "PSReadline" -ListAvailable).Version -ge [System.Version]"2.2.0") {
+    Set-PSReadLineOption -PredictionSource History
+    Set-PSReadLineOption -PredictionViewStyle ListView
+}
+else {
+    "Missing PSReadline 2.2.0 or newer, will not set options"
 }
 
 $script:MyOS = Get-MyOS