+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [5.0.0] - 2022-03-17
 ### Added
 - `Update-ProfileModule` also copies `profile.ps1` if module was updated
 
@@ -96,7 +98,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update-Profile now updates with local profile.ps1 from module, not internet
 
-[Unreleased]: https://github.com/codaamok/codaamok/compare/4.1.0..HEAD
+[Unreleased]: https://github.com/codaamok/codaamok/compare/5.0.0..HEAD
+[5.0.0]: https://github.com/codaamok/codaamok/compare/4.1.0..5.0.0
 [4.1.0]: https://github.com/codaamok/codaamok/compare/4.0.0..4.1.0
 [4.0.0]: https://github.com/codaamok/codaamok/compare/3.2.0..4.0.0
 [3.2.0]: https://github.com/codaamok/codaamok/compare/3.1.2..3.2.0